[storage] trivial: refactor SchemaIterator to its own mod
